CHROMIUM: drm/i915/pxp: cleanup pxp messages after signal interrupt
If pxp receive is interrupted by a signal, the message goes unhandled
until the next time a send/receive is called in which case the receive
is actually reading the previously unhandled message and the pxp
messages end up out of sync, causing no errors in the kernel but
userspace gets incorrect messages back.

Flag that the pxp receive was interrupted by signal and the next time we
do a send/receive, clear the previously unhandled message and read the
actual response.

Signed-off-by: Juston Li <juston.li@intel.com>

BUG=b:187211478
TEST=for i in {1..20}; do tast -verbose run <IP> video.PlayDRM*; done

Change-Id: I3551b1f74892fab9139fa315ee91bfc8e325b208
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3083693
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>