CHROMIUM: drm/i915/pxp: handle signal interrupt on pxp send
The following commit handled pxp messages getting interrupted when
waiting to receive a message, flagging to clean up on the next
send/receive.

commit 7b5415bdfa7bb0943b8fc10cb567b5346f48b429
Author: Alexander Usyskin <alexander.usyskin@intel.com>
Date:   Mon Aug 9 16:30:10 2021 -0700

    CHROMIUM: drm/i915/pxp: cleanup pxp messages after signal interrupt

It turns out that pxp send operations can also get interrupted by a
signal at a point when the message is in middle of being written.
In this case the message does end up getting sent bug we don't handle
EINTR for a pxp send so we end up with the same out of sync issue.

Handle EINTR here as well and indicate we need the same cleanup.

Signed-off-by: Juston Li <juston.li@intel.com>

BUG=b:187211478
TEST=tast -verbose run <IP> webrtc.GetDisplayMedia.monitor video.PlayDRM*

Change-Id: I7f7924cec0c027e380731b7abe73b4d045f3c808
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3098440
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>