CHROMIUM: swiotlb: Allow re-assigning same restricted DMA region
Restricted DMA regions, like other reserved memory regions, are
allocated in the device tree and instantiated early on. Also, the
mapping of these regions to devices is fixed in the device tree.
Thus when the restricted DMA region is reinitialized for a certain
device, it should always get the same region.

Allow reassigning the same restricted DMA region to a device when the
device already has a region associated with it. This gets rid of errors
during a re-probe of devices, such as due to unbind/rebind operations.

Note that this fix is specific to the v4 version of "swiotlb: Add
restricted DMA pool" in ChromeOS kernel v5.4. This was fixed in v5 and
above that was eventually merged upstream. This patch goes one step
further by emitting a warning if the new region does not match the
existing one.

Also, the mt7921e driver has some other issue, likely incomplete cleanup
at remove, that prevents rebind, which hinders testing of this patch.

BUG=b:149527406
BUG=b:183992356
TEST=unbind/rebind PCIe wifi on asurada should succeed

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: Iebdc25700231a011494a4eae572c125889e0ba97
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3099447
Reviewed-by: Tzung-Bi Shih <tzungbi@chromium.org>