UPSTREAM: coresight: syscfg: Initial coresight system configuration
Creates an system management API to allow complex configurations and
features to be programmed into a CoreSight infrastructure.

A feature is defined as a programming set for a device or class of
devices.

A configuration is a set of features across the system that are enabled
for a trace session.

The API will manage system wide configuration, and allow complex
programmed features to be added to individual device instances, and
provide for system wide configuration selection on trace capture
operations.

This patch creates the initial data object and the initial API for
loading configurations and features.

Link: https://lore.kernel.org/r/20210723165444.1048-2-mike.leach@linaro.org
Signed-off-by: Mike Leach <mike.leach@linaro.org>
Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Link: https://lore.kernel.org/r/20210818194022.379573-2-mathieu.poirier@linaro.org
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 85e2414c518a03a21dddd4bc88fec2723c5e1197)

BUG=b:185501434
TEST=CQ

Cq-Depend: chromium:3149834
Signed-off-by: Jian Cai <jiancai@google.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I73521f35d74c140d9638124ba520f1797c30f641
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3116698
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Manoj Gupta <manojgupta@chromium.org>