BACKPORT: FROMLIST: x86/mmu: Complete prefetch for trailing SPTEs for direct, legacy MMU
Make a final call to direct_pte_prefetch_many() if there are "trailing"
SPTEs to prefetch, i.e. SPTEs for GFNs following the faulting GFN.  The
call to direct_pte_prefetch_many() in the loop only handles the case
where there are !PRESENT SPTEs preceding a PRESENT SPTE.

E.g. if the faulting GFN is a multiple of 8 (the prefetch size) and all
SPTEs for the following GFNs are !PRESENT, the loop will terminate with
"start = sptep+1" and not prefetch any SPTEs.

Prefetching trailing SPTEs as intended can drastically reduce the number
of guest page faults, e.g. accessing the first byte of every 4kb page in
a 6gb chunk of virtual memory, in a VM with 8gb of preallocated memory,
the number of pf_fixed events observed in L0 drops from ~1.75M to <0.27M.

Note, this only affects memory that is backed by 4kb pages as KVM doesn't
prefetch when installing hugepages.  Shadow paging prefetching is not
affected as it does not batch the prefetches due to the need to process
the corresponding guest PTE.  The TDP MMU is not affected because it
doesn't have prefetching, yet...

Fixes: 957ed9effd80 ("KVM: MMU: prefetch ptes when intercepted guest #PF")
Signed-off-by: Sean Christopherson <seanjc@google.com>
Cc: Sergey Senozhatsky <senozhatsky@google.com>
Cc: Ben Gardon <bgardon@google.com>
Tested-by: Sergey Senozhatsky <senozhatsky@chromium.org>
(am from https://lore.kernel.org/patchwork/patch/1478751/)
(also found at https://lore.kernel.org/r/20210818235615.2047588-1-seanjc@google.com)

BUG=b:197183394
TEST=analyzed perf kvm stats report for VM boot up and
     apploading.MemoryTest

Signed-off-by: Sergey Senozhatsky <senozhatsky@chromium.org>
Change-Id: I4e30c761e239b9ef5c6029440f51faee95f317fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3106873
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>