CHROMIUM: config: Remove unused drivers from Qualcomm
We don't use these drivers on any chromeos devices so let's remove them
from the config. Leave them in the generic arm64 config so that we still
compile test them in case anyone has a device that may use them still.

This is generated by deleting the config and then running

    echo 'CONFIG_USB_DWC3_OF_SIMPLE=n' >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
    echo 'CONFIG_VIDEO_IR_I2C=n' >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
    ./chromeos/scripts/kernelconfig -F olddefconfig

BUG=b:189484858
TEST=tast run $dut kernel.Bloat; doesn't show these drivers

Change-Id: I8f2440bd52a6bf220ab84f1d2a30cf38aab92e0c
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3114587
Reviewed-by: Douglas Anderson <dianders@chromium.org>