CHROMIUM: Disable Remote Wakeup for RTL8822 USB
The parent of this patch enables RTL chips to remote wakeup in order
to let future gens of RTK chips to act as wakeup source. However,
it also enables RTL8822 to do so. Previously, RTL8822 remote wakeup
capability are disabled in order to save power.

This patch keeps remote wakeup for RTL8822 disabled in order to
preserve the condition.

BUG=b:178449733
TEST=verify BT keyboard cannot wake Ezkinil from suspend

Signed-off-by: Archie Pusaka <apusaka@chromium.org>
Change-Id: I8bdf506da4c7c575ab69502d44afa246c423d827
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3124177
Reviewed-by: Abhishek Pandit-Subedi <abhishekpandit@chromium.org>