CHROMIUM: config: Remove unused drivers from Qualcomm
We don't use the SMD driver on any chromeos devices so let's remove it
from the config. Leave it in the generic arm64 config so that we still
compile test it in case anyone has a device that may use the driver.

This is generated by deleting the config and then running

    echo 'CONFIG_RPMSG_QCOM_SMD=n' >> \
    chromeos/config/arm64/chromiumos-qualcomm.flavour.config
    ./chromeos/scripts/kernelconfig -F olddefconfig

BUG=b:189484858
TEST=tast run $dut kernel.Bloat; doesn't show these drivers

Disallow-Recycled-Builds: test-failures
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: If4bc4eb56175e653f26f62a03b3dac6b82f84a1d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3161599
Reviewed-by: Douglas Anderson <dianders@chromium.org>