CHROMIUM: iwl7000: mvm: check if ppag_allowed is valid before trying to match
The dmi_match() function also matches NULL, so if the platform doesn't
have a DMI_SYS_VENDOR defined and ppag_allowed is not set, it will
match, which is not what we intended.

Fix this by checking if ppag_allowed is defined before trying to match
with the list.

Change-Id: I880f9b8e7ef1e5b0b29b40af6ea4752ef313b577
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: f718d698b34d94dea2ec6aeb3a882ac398c39ac6