CHROMIUM: zram: Allow backing device to be assigned after init
I couldn't find a reason why we can't allow for the zram
backing device to be assigned after the device is initialized.

This change will allow the backing device to be assigned at any
point so long as it has not already been defined.

BUG=b:217186049
TEST=Tested on betty vm

Signed-off-by: Brian Geffon <bgeffon@google.com>
Change-Id: Ic1c2cbd7a2e74018d2ceb71b6fc885d01cae04f8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3575257
Reviewed-by: Yu Zhao <yuzhao@chromium.org>
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Tested-by: Brian Geffon <bgeffon@chromium.org>