CHROMIUM: usb: dwc3: Fixup non-wakeup suspend path
https://crrev.com/c/3192644 partially fixed some issues with [1], where
wakeup support was incorrectly assumed, but it missed the fact that [1]
was already not supporting non-wakeup modes correctly, as it failed to
balance dwc3_core_{init,exit}() (and therefore phy_*() states, etc.).

Rebalance things, by performing dwc3_core_exit() for host mode
non-wakeup (this is resolved in the latest [2]), and still suspend the
HS-PHY (*not* resolved in the latest [2]), because apparently that's
something that the Qualcomm driver expects.

[1] FROMLIST: usb: dwc3: core: Host wake up support from system suspend
    https://crrev.com/c/2352633
[2] https://lore.kernel.org/all/1624882097-23265-4-git-send-email-sanm@codeaurora.org/

BUG=b:163818407, b:196500382
TEST=scarlet, kevin, trogdor suspend/resume

Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: Ieb327e864f198a974284bea55ce6569200cf9b58
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3200149
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
(cherry picked from commit 82a52d6826a34291e914bc84676f8f371dacc57d)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3206495