BACKPORT: UPSTREAM: Bluetooth: Rename driver .prevent_wake to .wakeup
prevent_wake logic is backward since what it is really checking is
if the device may wakeup the system or not, not that it will prevent
the to be awaken.

Also looking on how other subsystems have the entry as power/wakeup
this also renames the force_prevent_wake to force_wakeup in vhci driver.

Signed-off-by: Luiz Augusto von Dentz <luiz.von.dentz@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit 4539ca67fe8edef34f522fd53da138e2ede13464)

Conflicts:
	drivers/bluetooth/hci_vhci.c

BUG=none
TEST=build

Signed-off-by: Zhengping Jiang <jiangzp@google.com>