CHROMIUM: iwl7000: don't pass actual WGDS revision number in table_revision
The FW API for PER_CHAIN_LIMIT_OFFSET_CMD is misleading.  The element
name is table_rev, but it shouldn't actually contain the table
revision number, but whether we should use the South Korea scheme or
not.

Fix the driver so that we only set this value to either 0 or 1.  It
will only be 1 (meaning South Korea) if the ACPI WGDS table revision
is 1.

Change-Id: I1fdc2c14577523fcffdfe8fb5902c2d8efde7e09
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: b73f8405382f672d2063ac144ed0399bc66164fd