BACKPORT: UPSTREAM: Bluetooth: mgmt: Fix Experimental Feature Changed event
This patch fixes the controller index in the Experimental Features
Changed event for the offload_codec and the quality_report features to
use the actual hdev index instead of non-controller index(0xffff) so the
client can receive the event and know which controller the event is for.

Fixes: ad93315183285 ("Bluetooth: Add offload feature under experimental flag")
Fixes: ae7d925b5c043 ("Bluetooth: Support the quality report events")
Signed-off-by: Tedd Ho-Jeong An <tedd.an@intel.com>
Signed-off-by: Marcel Holtmann <marcel@holtmann.org>
(cherry picked from commit b15bfa4df63529150df9ff0585675f728436e0c1)

Backport notes: skip codec related in the upstream.

Conflicts:
	net/bluetooth/mgmt.c

BUG=none
TEST=build

Signed-off-by: Zhengping Jiang <jiangzp@google.com>