UPSTREAM: drm/i915: Clean up PXP Kconfig info.
During the review I focused on stop the using of the "+"
to reference the newer platforms, but I forgot that we are
in a process of making things more clear and differentiate
graphics and display versions. So, let me to clean up this
a bit. Also, we don't need any version mentioned in the
config menu entry, only in the help.

Cc: Alan Previn <alan.previn.teres.alexis@intel.com>
Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20211015090916.82968-1-rodrigo.vivi@intel.com
(cherry picked from commit c974cf01b248c6f4220bfadd57cce74058453aea)

BUG=b:208454744
TEST=suspend_stress_test -c 1 during protected playback

Signed-off-by: Juston Li <juston.li@intel.com>
Change-Id: Ib6ee149390912961d9107325a01f099a3edbad03
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3309561
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Jeffrey Kardatzke <jkardatzke@google.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Tested-by: Jeffrey Kardatzke <jkardatzke@google.com>
Commit-Queue: Jeffrey Kardatzke <jkardatzke@google.com>