CHROMIUM: drm/panel-edp: Adapt to downstream DP AUX backlight driver
This is just like CL:3057149 ("CHROMIUM: drm/panel: atna33xc20: Adapt
to downstream DP AUX backlight driver") but for the generic edp-panel
driver.

NOTE: there are no panels that I'm aware of that use the DP AUX
backlight _and_ the edp-panel driver. However, without a patch like
this things won't compile.

This is done as a separate patch instead of being lumped into the
("BACKPORT: FROMGIT: drm/panel-edp: Split eDP panels out of
panel-simple") patch so that, in theory, it could be easily reverted
if we ever get the proper DP AUX backlight merged back to
chromeos-5.4.

BUG=b:179329054
TEST=Things compile again

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: Ic04bd5e210dd113ae36372132e61887d965ca862
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3227660
Reviewed-by: Sean Paul <seanpaul@chromium.org>