UPSTREAM: drm/msm/dp: Use the connector passed to dp_debug_get()
The debugfs code is provided an array of a single drm_connector. Then to
access the connector, the list of all connectors of the DRM device is
traversed and all non-DisplayPort connectors are skipped, to find the
one and only DisplayPort connector.

But as we move to support multiple DisplayPort controllers this will now
find multiple connectors and has no way to distinguish them.

Pass the single connector to dp_debug_get() and use this in the debugfs
functions instead, both to simplify the code and the support the
multiple instances.

Reviewed-by: Abhinav Kumar <abhinavk@codeaurora.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20211015232213.1839472-1-bjorn.andersson@linaro.org
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 899b2608d8d4b675cb54a026a27e28b4e1bcbf61)

BUG=b:200721581
TEST=emerge-trogdor chromeos-kernel-5_4

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: Ib4eb49b10fc041f26200af778cfd4aa36d24b180
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3406284
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>