UPSTREAM: drm/msm/dp: Remove global g_dp_display variable
As the Qualcomm DisplayPort driver only supports a single instance of
the driver the commonly used struct dp_display is kept in a global
variable. As we introduce additional instances this obviously doesn't
work.

Replace this with a combination of existing references to adjacent
objects and drvdata.

Reviewed-by: Abhinav Kumar <abhinavk@codeaurora.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Link: https://lore.kernel.org/r/20211016221843.2167329-2-bjorn.andersson@linaro.org
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit d624e50aa3c16283e2ee9f6644d9155697347ab5)

BUG=b:200721581
TEST=emerge-trogdor chromeos-kernel-5_4

Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I912e709638a4da466574f3fed767f02eb67648e0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3406286
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>