UPSTREAM: drm/msm/dp: Allow attaching a drm_panel
eDP panels might need some power sequencing and backlight management,
so make it possible to associate a drm_panel with an eDP instance and
prepare and enable the panel accordingly.

Now that we know which hardware instance is DP and which is eDP,
parser->parse() is passed the connector_type and the parser is limited
to only search for a panel in the eDP case.

Signed-off-by: Bjorn Andersson <bjorn.andersson@linaro.org>
Reviewed-by: Stephen Boyd <swboyd@chromium.org>
Link: https://lore.kernel.org/r/20211016221843.2167329-5-bjorn.andersson@linaro.org
Signed-off-by: Rob Clark <robdclark@chromium.org>
(cherry picked from commit 4b296d15b355a866504e60af85ebb72b47873206)

BUG=b:200721581
TEST=emerge-trogdor chromeos-kernel-5_4

Cq-Depend: chromium:3406291
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
Change-Id: I91d73aed5efa31fbb652ff5e0644db991db04eea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3406289
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>