CHROMIUM: iwl7000: mvm: align locking in D3 test debugfs
Since commit a05829a7222e ("cfg80211: avoid holding the RTNL when
calling the driver") we're not only holding the RTNL when going
in and out of suspend, but also the wiphy->mtx. Add that to the
D3 test debugfs in iwlwifi since it's required for various calls
to mac80211.

Change-Id: Ib73bf787ab4d83581de20eb89b1f8dbfcaaad0e3
Fixes: a05829a7222e ("cfg80211: avoid holding the RTNL when calling the driver")
Signed-off-by: Johannes Berg <johannes.berg@intel.com>
iwl7000-tree: 28fda7e96ef440707f111038ba13b8c83ec166b2