CHROMIUM: ASoC: Intel: Skylake: Set dsp cores off during shutdown
When the cores.usage_count is equal to 0, driver puts dsp cores to
sleep.  The issue happens when cores.usage_count is not equal to 0
and dsp core remains ON even when the system goes to shutdown.

Removing the dependency of usage_count by forcing to disable
the dsp core.

BUG=b:201547645
TEST=stress test 1000+ shutdown cycles on akali360

Signed-off-by: Mohit Sinha <mohit.sinha@intel.com>
Signed-off-by: Zhang Yanmin <yanmin.zhang@intel.com>
Signed-off-by: Matt Davis <mattedavis@google.com>
Change-Id: Ie3c408eedf2f4e51a051dfbf48dfca51b2db8eb3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3291690
Reviewed-by: Curtis Malainey <cujomalainey@chromium.org>