BACKPORT: FROMLIST: drm/msm/a6xx: Skip crashdumper state if GPU needs_hw_init
I am seeing some crash logs which imply that we are trying to use
crashdumper hw to read back GPU state when the GPU isn't initialized.
This doesn't go well (for example, GPU could be in 32b address mode
and ignoring the upper bits of buffer that it is trying to dump state
to).

I'm not *quite* sure how we get into this state in the first place,
but lets not make a bad situation worse by triggering iova fault
crashes.

While we're at it, also add the information about whether the GPU is
initialized to the devcore dump to make this easier to see in the
logs (which makes the WARN_ON() redundant and even harmful because
it fills up the small bit of dmesg we get with the crash report).

Signed-off-by: Rob Clark <robdclark@chromium.org>
(am from https://patchwork.freedesktop.org/patch/466167/)
(also found at https://lore.kernel.org/r/20211209193118.1163248-1-robdclark@gmail.com)

Conflicts:
   drivers/gpu/drm/msm/adreno/adreno_gpu.c

BUG=b:204345099
TEST=boot homestar

Change-Id: Ib51f08b36b94efd27eb1be50f02519042b0a5336
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3328094
Tested-by: Rob Clark <robdclark@chromium.org>
Auto-Submit: Rob Clark <robdclark@chromium.org>
Reviewed-by: Kristian Kristensen <hoegsberg@chromium.org>
Commit-Queue: Kristian Kristensen <hoegsberg@chromium.org>