CHROMIUM: arm64: dts: sc7180: trogdor: Add compatible strings for RTS5414 USB hub
trogdor/stronbad devices can be stuffed with either the RTS5411 or
RTS5414 USB hub. So far the device tree only contains compatible
strings for the RTS5411, add the missing RTS5414 ones.

It is unusual to specify compatible strings for multiple devices in
a single DT node. In practice none of the compatible strings is
actually used, the USB core identifies the driver based on the VID:PID
pairs that are reported by the device.

BUG=b:212740825
TEST=manual on homestar

Change-Id: Ib8f4de7d63ce6b7334bca1fee7fe02e889de6afc
Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3366505
Reviewed-by: Douglas Anderson <dianders@chromium.org>