UPSTREAM: net: ipa: fix atomic update in ipa_endpoint_replenish()
In ipa_endpoint_replenish(), if an error occurs when attempting to
replenish a receive buffer, we just quit and try again later.  In
that case we increment the backlog count to reflect that the attempt
was unsuccessful.  Then, if the add_one flag was true we increment
the backlog again.

This second increment is not included in the backlog local variable
though, and its value determines whether delayed work should be
scheduled.  This is a bug.

Fix this by determining whether 1 or 2 should be added to the
backlog before adding it in a atomic_add_return() call.

Reviewed-by: Matthias Kaehlcke <mka@chromium.org>
Fixes: 84f9bd12d46db ("soc: qcom: ipa: IPA endpoints")
Signed-off-by: Alex Elder <elder@linaro.org>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6c0e3b5ce94947b311348c367db9e11dcb2ccc93)

BUG=b:200593656
TEST=emerge-herobrine chromeos-kernel-5_15

Signed-off-by: Matthias Kaehlcke <mka@chromium.org>
Change-Id: I7648efa2011764bb56e39ce635479901b7d7dfc1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3415221
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Douglas Anderson <dianders@chromium.org>