FIXUP: BACKPORT: FROMGIT: Bluetooth: mgmt: Make use of mgmt_send_event_skb in MGMT_EV_DEVICE_FOUND
There was conflict while cherry-picking from upstream between these
patches:
Fixes: 756111e95c ("BACKPORT: FROMGIT: Bluetooth: mgmt: Make use of mgmt_send_event_skb in MGMT_EV_DEVICE_FOUND")
Fixes: 066579b8aa ("FROMGIT: Bluetooth: Send device found event on name resolve failure")

The order of merging in upstream is 2 1, however it's merged as 1 2 in
Chromium tree. Since patch 2 depends on patch 1, it was CP'd as BACKPORT
but single line was omitted, which caused the skb to be shorter than
required. This patch fixes the problem. FIXUP was chosen as solution for
easier tracking as the same patches were merged to 4.14 4.19 5,4 and 5.10.
This patch will land on all of those CHROMIUM lines.  For 5.15 patches will
be merged with the correct order.

BUG=b:212279838,b:201527418
TEST=Bug reproduced on HID devices like Bluetooth Mouse M336/M337/M535
     Build kernel and update on DUT
     Remove cache: rm -f /var/lib/bluetooth/<HostAddr>/cache/*
     Restart DUT (important)
     Run bluetoothctl
     # scan on
     # info <DeviceAddr>
     Verification passed if Name and Alias fields return
         proper name of device
     Repeat the cache cleaning stage and DUT reset (important)
     Login to OS
     Verify HID devices can be found by system UI

Signed-off-by: Angela Czubak <acz@semihalf.com>
Signed-off-by: Marek Maslanka <mm@semihalf.com>
Signed-off-by: Radoslaw Biernacki <rad@semihalf.com>
Signed-off-by: Archie Pusaka <apusaka@chromium.org>
Change-Id: I971878a7b7c033f470511d130d5f633c059e873a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3373028
Reviewed-by: Bob Brandt <brndt@google.com>
Reviewed-by: Guenter Roeck <groeck@chromium.org>
Commit-Queue: Radoslaw Biernacki <biernacki@google.com>
Tested-by: Radoslaw Biernacki <biernacki@google.com>
(cherry picked from commit a1c4b1c8329913b7769106c2c356a3e5ade8bc8f)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3388572
Reviewed-by: Radoslaw Biernacki <biernacki@google.com>