CHROMIUM: media: mtk-vcodec: Adapt stateless VP9 decoder to new final uAPI
With the migration of the VP9 stateless decoder uAPI to the finalized
version, the mtk-vcodec driver needs to be migrated as well. However
this part of the driver is not fully upstreamed yet, and the current
downstream version is also very different from the latest upstream
mailing list one.

Instead of trying to back out the whole driver and backport the latest
stuff, causing a lot of code churn and most likely breaking things in
the process, it is much easier to just update the driver to match the
new uAPI. The changes are the same as the ones that were made to the
upstream patches.

Upon uprev for cherry, the whole stack will be thrown away, and
hopefully upstream will already be in a usable state.

BUG=b:214332361
BUG=b:226978629
TEST=Build for asurada (with next revert patch)

Signed-off-by: Chen-Yu Tsai <wenst@chromium.org>
Change-Id: I6ad3fdf30192b8cd4b8563f5f6a31bbdc143bff7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3449951
Reviewed-by: Miguel Casas <mcasas@google.com>
Reviewed-by: Brian Norris <briannorris@chromium.org>
Commit-Queue: Brian Norris <briannorris@chromium.org>
(cherry picked from commit 1a14b037201726a6eb3a24bc449a2cdc3fded0d2)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3576174