CHROMIUM: config: Introduce config family for kernel build
We have couple of different builds that could be based on the same
kernel source but with different config, for eg: default chromeos build,
manatee, termina etc. Instead of having separate source folders or
maintaining single CHROMEOS_KERNEL_CONFIG, introduce the concept of
family with in the same source tree. There will be one config folder
per family in the source tree which maintains the split config.

Re-organize the folder structure for accomodating the concept of family.
Later patches will update the scripts to understand the new folder
structure.

BUG=None
TEST=None

Signed-off-by: Vineeth Pillai <vineethrp@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3398672
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Signed-off-by: Brian Norris <briannorris@chromium.org>
Change-Id: I19638468b15d20ed5dd97ff08b56379837ec4826
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3617051
Reviewed-by: Douglas Anderson <dianders@chromium.org>