BACKPORT: FROMGIT: drm: Plumb debugfs_init through to panels
We'd like panels to be able to add things to debugfs underneath the
connector's directory. Let's plumb it through. A panel will be able to
put things in a "panel" directory under the connector's
directory. Note that debugfs is not ABI and so it's always possible
that the location that the panel gets for its debugfs could change in
the future.

NOTE: this currently only works if you're using a modern
architecture. Specifically the plumbing relies on _both_
drm_bridge_connector and drm_panel_bridge. If you're not using one or
both of these things then things won't be plumbed through.

As a side effect of this change, drm_bridges can also get callbacks to
put stuff underneath the connector's debugfs directory. At the moment
all bridges in the chain have their debugfs_init() called with the
connector's root directory.

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Javier Martinez Canillas <javierm@redhat.com>
Reviewed-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20220204161245.v2.2.Ib0bd5346135cbb0b63006b69b61d4c8af6484740@changeid

(cherry picked from commit 2509969a9862b522d2208e8663057fb227556687
 git://anongit.freedesktop.org/drm/drm-misc drm-misc-next)

Conflicts:
   drivers/gpu/drm/drm_debugfs.c
   include/drm/drm_connector.h
Both are simple context conflicts and trivial to resolve. We lack
commit 41752663b410 ("drm/debug: Expose connector VRR monitor range
via debugfs") and commit 72ad49682dde ("drm/connector: Add support for
out-of-band hotplug notification (v3)") and picking those patches
would be more involved.

BUG=b:215432805
TEST=grep "" /sys/kernel/debug/dri/*/eDP*/panel/detected_panel

Change-Id: Ib0bd5346135cbb0b63006b69b61d4c8af6484740
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3470514
Tested-by: Douglas Anderson <dianders@chromium.org>
Reviewed-by: Drew Davenport <ddavenport@chromium.org>
Commit-Queue: Douglas Anderson <dianders@chromium.org>