CHROMIUM: iwl7000: mvm: Don't fail if PPAG isn't supported
When we're copying the PPAG table into the cmd structure we're failing
if the table doesn't exist in ACPI or is invalid, or if the FW doesn't
support PPAG setting etc.

This is wrong because those are valid scenarios.  Fix this by not
failing in those cases.

Change-Id: I6a9c65149c2c3c11337f3a802dff22f514a3a436
Signed-off-by: Miri Korenblit <miriam.rachel.korenblit@intel.com>
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
iwl7000-tree: bfd21f232687d9ec26c32de3cf554b3bf03cbdb3