CHROMIUM: bluetooth: Support userspace SCO
In order to support userspace SCO, we need to expose a few commands:
* get_sco_codec_capabilities
* notify_sco_connection_state

get_sco_codec_capabilities notifies userspace what capabilities the
current controller has for a given set of codecs. It also returns the
vendor specific data that needs to be sent with HCI_CONFIGURE_DATAPATH.
Currently, the offload path is not supported.

notify_sco_connection_state notifies the kernel (and specifically the
driver) when a sco connection is established or removed. This allows the
driver to respond (such as by switching altsettings in the USB case).

BUG=b:223685328
TEST=emerge-zork chromeos-kernel-5_4

Change-Id: I9cd56b30015aeadab086d6b93fd1c7fd6da5e058
Signed-off-by: Abhishek Pandit-Subedi <abhishekpandit@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3564861
Reviewed-by: Ying Hsu <yinghsu@chromium.org>