FROMLIST: dt-bindings: iio: sx9324: Add CS idle/sleep mode
We need to configure the sleep and idle mode of the CS pins for this
device. Add a DT property to do this so pins are in a proper state
during sleep (either hi-z, grounded, or pulled up to vdd).

Cc: Gwendal Grignou <gwendal@chromium.org>
Signed-off-by: Stephen Boyd <swboyd@chromium.org>
(am from https://lore.kernel.org/r/20220503004156.3559940-2-swboyd@chromium.org)

BUG=b:203837652
TEST=emerge-trogdor chromeos-kernel-5_4; Boot on pazquel

Change-Id: I853ccfae65676a83a6d320c9dbe9c0d56b8c7ca8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3620795
Reviewed-by: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Stephen Boyd <swboyd@chromium.org>
Commit-Queue: Stephen Boyd <swboyd@chromium.org>