CHROMIUM: mtk-vcodec: fix data_offset buffer overrun
Thanks to Nicolas Dufresne for pointing out both the bug and the fix.

When we validate that the image size does not exceed the size of the
plane buffer, we are currently neglecting to count the data_offset in
our calculation. Most drivers simple ignore data_offset, but the MTK
vcodec driver has been hacked specifically to support it for ARC++
encoding. See cr/2721869.

Note that this CL will need to be ported for 5.10 and 5.15 as well.

BUG=b:235392268
TEST=Kernel builds and runs on Asurada
Suggested-by: Nicolas Dufresne <nicolas.dufresne@collabora.com>
Signed-off-by: Justin Green <greenjustin@google.com>

Change-Id: I00d51a514861f35b50d5882792380ce5a94d464d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3696583
Tested-by: Justin Green <greenjustin@google.com>
Commit-Queue: Justin Green <greenjustin@google.com>
Reviewed-by: Miguel Casas <mcasas@google.com>