FROMLIST: drm/edid: ignore the CEA modes not defined in CEA-861-D
This is a workaround for HDMI 1.4 sink which has a CEA mode with higher vic
than what is defined in CEA-861-D.

As an example, a HDMI 1.4 sink has the video format 2560x1080p to be
displayed and the video format is indicated by both SVD (with vic 90 and
picture aspect ratio 64:27) and DTD.  When connecting to such sink,
source can't output the video format in SVD because an error is returned by
drm_hdmi_avi_infoframe_from_display_mode(), which can't fill the infoframe
with picture aspect ratio 64:27 and the vic, which is originally 90 and is
changed to 0 by drm_mode_cea_vic().

To work around it, do not set the vic 0 so the corresponding mode may be
accepted in drm_hdmi_avi_infoframe_from_display_mode() and be dispalyed.

Cc: Ville Syrjälä <ville.syrjala@linux.intel.com>
Cc: Jani Nikula <jani.nikula@linux.intel.com>
Cc: Wayne Lin <waynelin@amd.com>
Cc: Lee Shawn C <shawn.c.lee@intel.com>
Signed-off-by: William Tseng <william.tseng@intel.com>
(am from https://patchwork.freedesktop.org/patch/489255)

Modify: Remove const parameter from drm_mode_cea_vic for fix build error

BUG=b:230691248,b:236781087
TEST=Verify on Project Cret
Signed-off-by: annya_chen <annya_chen@compal.corp-partner.google.com>
Change-Id: I1e2cfd1e45d934028642ed2630401ba224aec6a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3682725
Tested-by: AJ Cheng <aj_cheng@compal.corp-partner.google.com>
Reviewed-by: Sean Paul <sean@poorly.run>
Reviewed-by: Evan Green <evgreen@chromium.org>
Commit-Queue: Shou-Chieh Hsu <shouchieh@chromium.org>