FROMGIT: drm/msm/adreno: Defer enabling runpm until hw_init()
To avoid preventing the display from coming up before the rootfs is
mounted, without resorting to packing fw in the initrd, the GPU has
this limbo state where the device is probed, but we aren't ready to
start sending commands to it.  This is particularly problematic for
a6xx, since the GMU (which requires fw to be loaded) is the one that
is controlling the power/clk/icc votes.

So defer enabling runpm until we are ready to call gpu->hw_init(),
as that is a point where we know we have all the needed fw and are
ready to start sending commands to the coproc's.

Signed-off-by: Rob Clark <robdclark@chromium.org>
Patchwork: https://patchwork.freedesktop.org/patch/489337/
Link: https://lore.kernel.org/r/20220613182036.2567963-1-robdclark@gmail.com
(cherry picked from commit 4b18299b33655fa9672b774b6df774dc03d6aee8
 https://gitlab.freedesktop.org/drm/msm.git msm-next)

BUG=b:230129261, b:209880254
TEST=Build and boot

Signed-off-by: Douglas Anderson <dianders@chromium.org>
Change-Id: I047dc21d3a13ea729c837c09c168ef0a10338fea
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3756002
Reviewed-by: Rob Clark <robdclark@chromium.org>
Reviewed-by: Sean Paul <sean@poorly.run>