renoir_defconfig: Re-enable workaround for erratum 1024718
This is relevant on Cortex-A55s, so enable it.

Note that dmesg will say that hardware dirty bit management has been
enabled when it gets enabled for at least one of the cores in the CPU,
which can be mislead people into thinking that the workaround wasn't applied
when it actually was.

Signed-off-by: Kazuki Hashimoto <kazukih@tuta.io>