CHROMIUM: gpu: mali: mediatek: Apply refactored platform code
The refactoring was originated in 5.15 kernel and was ported to the Mali
private repo in CL:*4843730. Refer to that for the full commit list.

Instead of having a commit squash similar to CL:*4843730, I rsync'd the
code from the private repo so that we can also get some new Mali r32p0
updates without any conflict:

  BASE=~/chromiumos/src/partner_private/mali-ddk-valhall
  DIR=drivers/gpu/arm/valhall/platform/mediatek/
  rsync -azvhP $BASE/kernel/$DIR ./$DIR

BUG=b:227544156
TEST=build Asurada kernel
     boot DUT to login screen
     run graphics_WebGLAquarium test

Change-Id: Ifd2a26d9998e72ab7bf3ef2225573a9a224c0030
Signed-off-by: Fei Shao <fshao@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3726589
Reviewed-by: Chen-Yu Tsai <wenst@chromium.org>