CHROMIUM: Bluetooth: Ignore global Connectable setting in hci_sync
Existing logic enforces a "connectable" advertisement if the global
Connectable setting is active, even if the advertisement requested was
non-connectable. This can cause a case where the client requires a
random address, but kernel will ignore that request. This patch removes
the check for the global setting, to ensure we provide a random address
if the client did not specify Connectability in their advertisement
request.

Specifically, Intel-AX201 and WCN3991 devices sometimes fail on
adv_nearby_test. The global Connectable setting makes this advertisement
public (not good), and then future discovery requests fail because the
random address update is skipped since adv is enabled.

Change was tested by setting Connectable on and run the adv_nearby_test
test, while the whole AdvHealth batch is tested as well to ensure that
it doesn't break any basic functionalities.

BUG=b:237249568
TEST=Set Connectable on and run adv_nearby_test test

Signed-off-by: Hsin-chen Chuang <chharry@chromium.org>
Change-Id: I2e2734f8b65f70515ec52190fecec9dee1b0937c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3740264
Reviewed-by: Miao-chen Chou <mcchou@chromium.org>
Reviewed-by: Archie Pusaka <apusaka@chromium.org>
Commit-Queue: Hsin-chen Chuang <chharry@google.com>
Tested-by: Hsin-chen Chuang <chharry@google.com>