CHROMIUM: Bluetooth: btusb: Workaround for spotty SCO quality
When streaming HFP, about once every 5 minutes a brief pause in audio
can be heard. The root cause is still under investigation, meanwhile
this CL is a workaround to mitigate that issue. Once the root cause
is fixed, this needs to be reverted, tracked in b/238575342.

BUG=b:233895851
TEST=Test streaming HFP for 1 hour on Dewatt, there is no noticeable
     gap in audio. Syslog doesn't complain about wrong SCO handle.

Signed-off-by: Hilda Wu <hildawu@realtek.com>
Signed-off-by: Archie Pusaka <apusaka@chromium.org>

Change-Id: Ibd1abe54e7b7b8bfe19a99bd6c36d59533a3db84
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3764210
Commit-Queue: Archie Pusaka <apusaka@chromium.org>
Tested-by: Archie Pusaka <apusaka@chromium.org>
Reviewed-by: Ying Hsu <yinghsu@chromium.org>