CHROMIUM: zram: Don't allow rw_page when not synchronous
Because zram on 5.4 kernels does not support changing the device
ops this means we will always have a rw_page method even when the device
is not synchronous. This change will return -EOPNOTSUPP if a user
invokes rw_page when the device is not synchronous.

Signed-off-by: Brian Geffon <bgeffon@chromium.org>

BUG=b:239824885
TESTED=betty vm w/ 5.4 kernel

Change-Id: I610181bd7fe772a87a5504bd4fbc19ecd7a0b465
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3795296
Commit-Queue: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
Tested-by: Brian Geffon <bgeffon@chromium.org>