CHROMIUM: config: Enable CONFIG_UNICODE
After the planned migration from virtio-fs to virtio-blk for Android
/data directory (go/arcvm-virtio-blk-data), the content of Android /data
will no longer be available at /home/root/<hash>/android-data/data on
the host side.
Currently, some tast tests access the directory from the host side with
this path. In order to accommodate such tests with the virtio-blk /data
migration, we are going to mount the virtio-blk disk image of Android
/data direcory on the host side at /home/root/<hash>/android-data/data
in these tests. Since the file system for Android /data uses casefold
feature and cannot be mounted on the host without CONFIG_UNICODE, we
are enabling it on all host kernel versions that support this config.

BUG=b:233565855
TEST=Check that ARCVM guest disk image (which has casefold feature) can
TEST=be mounted on the host.
TEST=Run `chromeos/scripts/kernelconfig checkconfig`.

Signed-off-by: Momoko Hattori <momohatt@google.com>
Change-Id: I017571804ce080ced17c4be891b7331f8392f111
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/3816192
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Sangwhan Moon <sxm@google.com>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>